Remove function!
Added the removal function on the Item Frame breake wich means that I now save the Armor Stands UUID in the lore of the display item.

I also tried to fix an Issue, wich happens at the creation of the shop... couldn't fix it.

And I fixed a aestetic issue in the config (added the line under the p in WaterDropMC)